// prisma/generated/zod/inputTypeSchemas/OAuthTokenUpdateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { NullableStringFieldUpdateOperationsInputSchema } from './NullableStringFieldUpdateOperationsInputSchema';
import { NullableDateTimeFieldUpdateOperationsInputSchema } from './NullableDateTimeFieldUpdateOperationsInputSchema';
import { NullableJsonNullValueInputSchema } from './NullableJsonNullValueInputSchema';
import { InputJsonValueSchema } from './InputJsonValueSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { AppUpdateOneRequiredWithoutOAuthTokenNestedInputSchema } from './AppUpdateOneRequiredWithoutOAuthTokenNestedInputSchema';
import { MerchantUpdateOneRequiredWithoutOAuthTokenNestedInputSchema } from './MerchantUpdateOneRequiredWithoutOAuthTokenNestedInputSchema';

export const OAuthTokenUpdateInputSchema: z.ZodType<Prisma.OAuthTokenUpdateInput> = z.object({
  id: z.union([ z.string().cuid(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  accessToken: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  refreshToken: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  expiresAt: z.union([ z.coerce.date(),z.lazy(() => NullableDateTimeFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  scopes: z.union([ z.lazy(() => NullableJsonNullValueInputSchema),InputJsonValueSchema ]).optional(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  app: z.lazy(() => AppUpdateOneRequiredWithoutOAuthTokenNestedInputSchema).optional(),
  merchant: z.lazy(() => MerchantUpdateOneRequiredWithoutOAuthTokenNestedInputSchema).optional()
}).strict();

export default OAuthTokenUpdateInputSchema;